Experimental Filter Code
This is some experimental DCM filter code. The results are better, but
still not idea. This is with no tuning/

Part of the problem seems to lie with the magnetometer Yaw calculation,
and the physical relative orientation of the free-floating magnetometer
and board mounted accelerometer. I am learning more about the theory
before i proceed.